requirements.txt
Django==4.1
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
murmurhash==1.0.8
numpy==1.23.2
packaging==21.3
pathy==0.6.2
preshed==3.0.7
pydantic==1.9.2
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.2
srsly==2.4.4
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
wasabi==0.10.1
